| Q             | A
| ------------- | ---
| Branch?       | 6.3
| Bug fix?      | no
| New feature?  | no
| Deprecations? | no
| Tickets       | 
| License       | MIT
| Doc PR        | 

When you are using a lot of voters, you need better caching. With inspiration from https://github.com/symfony/symfony/pull/43066 by @jderusse I managed to squeeze some more performance from the `AccessDecisionManager`.

I do about 7000 calls to `AccessDecisionManager::getVoters()`. It takes me 369ms. 
With this PR, it takes about 30ms. 

The trick is that I don't loop over all Voters all the time. If I seen the object and attribute before then I just returns the voters I previously calculated. 

------------

### Performance:

When do benchmark on an application with 40 voters I got: 

| Metric | From | To |
|------- |------|----|
| Wall time | 1150ms | 859 ms
| I/O Wait | 135 ms | 154 ms
| CPU | 1020ms | 705 ms
| Memory | 11.8 MB | 12.9 MB

-------------

I would love a few more set of eyes on this PR.