@Nyholm 2023-03-11
I spent 3 hours last night considering if I should make this PR or not. 
I had this feeling that something wasn't right.. This fix was "too simple" for @jderusse or anyone else to miss. 

I slept on it but still couldn't understand the why I got so massive performance gain... It turns out that I had xdebug on while debugging 🤦🏽 

When I re-ran the profiles it turns out they are pretty much the same. I dont think this complexity is worth the almost non existent gain.